Added stress domain parameter to truss and beam domains to that stress plot variable can be simplified.

+++ FEBioMech/FEExplicitSolidSolver.cpp
@@ -715,7 +715,7 @@ void FEExplicitSolidSolver::Serialize(DumpStream& ar)
 
 		if (ar.IsSaving())
 		{
-			int N = m_data.size();
+			int N = (int)m_data.size();
 			ar << N;
 			for (int i = 0; i < N; ++i)
 			{
